refactor(*): lotto 관련 클래스들을 domain package로 이동
lotto 관련 클래스들을 domain package로 이동시켰습니다.
