+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.Lotto;
-
 public interface LottoGenerator {
     Lotto generate();
 }
