+++ src/main/java/lotto/domain/LottoNumbersInfo.java
@@ -1,9 +1,9 @@
-package lotto.dto;
+package lotto.domain;
 
 import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.List;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.message.LottoResult;
 import lotto.utils.LottoUtil;
