+++ src/main/java/lotto/domain/LottoPurchase.java
@@ -2,7 +2,6 @@ package lotto.domain;
 
 import java.math.BigDecimal;
 import java.util.HashMap;
-import lotto.dto.LottoNumbersInfo;
 import lotto.message.LottoResult;
 
 public class LottoPurchase {
