+++ src/test/java/lotto/domain/LottoNumbersInfoTest.java
@@ -6,8 +6,6 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
-import lotto.Lotto;
-import lotto.dto.LottoNumbersInfo;
 import lotto.message.LottoResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
