+++ src/test/java/lotto/utils/LottoUtilTest.java
@@ -1,0 +1,8 @@
+package lotto.utils;
+
+import org.junit.jupiter.params.ParameterizedTest;
+
+public class LottoUtilTest {
+
+
+}
